- Get rid of AbstractTcp/UdpTransport classes (not needed yet, can be added back later if necessary) - Add deprecated Tcp/UdpTransport classes for backwards compatibility
